Enforce types through global model

+++ profsea/components/core/state.py
@@ -21,7 +21,7 @@ class ClimateState:
     nyr: int
     nt: int
     num_members: int
-    dtype: np.dtype = np.float64
+    dtype: np.dtype = np.float32
 
 
 @dataclass
@@ -35,7 +35,7 @@ class SpatialState:
     grid_interpolation: str
     output_percentiles: list[int] | np.ndarray
     baseline_yrs: tuple[int, int]
-    dtype: np.dtype = np.float64
+    dtype: np.dtype = np.float32
 
 
 @dataclass
@@ -49,4 +49,4 @@ class LocalState:
     interpolation_method: str
     output_percentiles: list[int] | np.ndarray
     baseline_yrs: tuple[int, int]
-    dtype: np.dtype = np.float64
+    dtype: np.dtype = np.float32
